pyproject.toml
[tool.poetry]
name = "api-bancaria"
version = "0.1.0"
description = "\"API Bancária Assíncrona com FastAPI\""
authors = ["Izairton Oliveira de Vasconcelos <iovascon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.0"
databases = "^0.9.0"
pydantic = "^2.9.2"
python-jose = "^3.3.0"
pydantic-settings = "^2.5.2"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
